Applied project defined formatting to all files

+++ proj/src/devices/timer/timer.h
@@ -3,20 +3,17 @@
 
 #include <lcom/lcf.h>
 
-
 /** @defgroup timer timer
- * 
+ *
  * Additional functions used for programming the timer
  */
 
-
 /**
  * @brief get the current counter valur for the timer
- * 
+ *
  * @return the current counter value
  */
-unsigned long (get_timer_get_counter)();
-
+unsigned long(get_timer_get_counter)();
 
 /**@}*/
